Add files via upload
These two GUIs have updates that include _unwrap as well as the updates to the manifest matcher. now when the Barcodes are scanned there is a user visible manifest that allows for the user to see what has been scanned and what needs to be scanned still. The scans that are registered now will be able to 'Greyed' out as a visual indicator.